+++ frontend/src/features/room/order/components/PlayerMenu.styled.ts
@@ -1,0 +1,13 @@
+import styled from '@emotion/styled';
+
+export const OrderList = styled.div`
+  height: calc(100% - 4rem);
+  overflow-y: scroll;
+`;
+
+export const OrderItem = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 0.75rem 0;
+`;
